refactor: test 파일들을 domain directory로 이동
